requirements.txt
openai
pdfplumber
pdf2image
label-studio
numpy
scikit-learn
flask
config
pytesseract
requests
pillow